requirements.txt
google-generativeai>=0.8.3
pymongo[srv]>=4.6.1
fastapi>=0.111.0
uvicorn>=0.30.1
langgraph>=0.0.65
tavily-python>=0.3.3
python-dotenv>=1.0.1
pydantic>=2.7.4
langchain-core>=0.2.10
langchain-text-splitters>=0.2.2
pymupdf>=1.24.5
